suggested changesby the code reviewer  have been incorporated
